Update Multiple-Regression.R
The original regression analysis predicted each participant's shift score. This revision added code for an alternative regression analysis predicting the average PR score of each ideation session as a function of both the problem context and the framing. Requires reorganizing the data frame so each ideation session is a row in the data frame.